+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:google_maps_widget/google_maps_widget.dart';
 
@@ -6,48 +8,101 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
+  // can create a controller, and call methods to update source loc,
+  // destination loc, interact with the google maps controller to
+  // show/hide markers programmatically etc.
+  final mapsWidgetController = GlobalKey<GoogleMapsWidgetState>();
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: SafeArea(
         child: Scaffold(
-          body: GoogleMapsWidget(
-            apiKey: "YOUR KEY HERE",
-            sourceLatLng: LatLng(40.484000837597925, -3.369978368282318),
-            destinationLatLng: LatLng(40.48017307700204, -3.3618026599287987),
+          body: Column(
+            children: [
+              Expanded(
+                child: GoogleMapsWidget(
+                  apiKey: "YOUR GOOGLE MAPS API KEY HERE",
+                  key: mapsWidgetController,
+                  sourceLatLng: LatLng(
+                    40.484000837597925,
+                    -3.369978368282318,
+                  ),
+                  destinationLatLng: LatLng(
+                    40.48017307700204,
+                    -3.3618026599287987,
+                  ),
 
-            ///////////////////////////////////////////////////////
-            //////////////    OPTIONAL PARAMETERS    //////////////
-            ///////////////////////////////////////////////////////
+                  ///////////////////////////////////////////////////////
+                  //////////////    OPTIONAL PARAMETERS    //////////////
+                  ///////////////////////////////////////////////////////
 
-            routeWidth: 2,
-            sourceMarkerIconInfo: MarkerIconInfo(
-              assetPath: "assets/images/house-marker-icon.png",
-            ),
-            destinationMarkerIconInfo: MarkerIconInfo(
-              assetPath: "assets/images/restaurant-marker-icon.png",
-            ),
-            driverMarkerIconInfo: MarkerIconInfo(
-              assetPath: "assets/images/driver-marker-icon.png",
-              assetMarkerSize: Size.square(125),
-              rotation: 90,
-            ),
-            updatePolylinesOnDriverLocUpdate: true,
-            // mock stream
-            driverCoordinatesStream: Stream.periodic(
-              Duration(milliseconds: 500),
-              (i) => LatLng(
-                40.47747872288886 + i / 10000,
-                -3.368043154478073 - i / 10000,
+                  routeWidth: 2,
+                  sourceMarkerIconInfo: MarkerIconInfo(
+                    assetPath: "assets/images/house-marker-icon.png",
+                  ),
+                  destinationMarkerIconInfo: MarkerIconInfo(
+                    assetPath: "assets/images/restaurant-marker-icon.png",
+                  ),
+                  driverMarkerIconInfo: MarkerIconInfo(
+                    assetPath: "assets/images/driver-marker-icon.png",
+                    assetMarkerSize: Size.square(125),
+                    rotation: 90,
+                  ),
+                  updatePolylinesOnDriverLocUpdate: true,
+                  // mock stream
+                  driverCoordinatesStream: Stream.periodic(
+                    Duration(milliseconds: 500),
+                    (i) => LatLng(
+                      40.47747872288886 + i / 10000,
+                      -3.368043154478073 - i / 10000,
+                    ),
+                  ),
+                  sourceName: "This is source name",
+                  driverName: "Alex",
+                  onTapDriverMarker: (currentLocation) {
+                    print("Driver is currently at $currentLocation");
+                  },
+                  totalTimeCallback: (time) => print(time),
+                  totalDistanceCallback: (distance) => print(distance),
+                ),
               ),
-            ),
-            sourceName: "This is source name",
-            driverName: "Alex",
-            onTapDriverMarker: (currentLocation) {
-              print("Driver is currently at $currentLocation");
-            },
-            totalTimeCallback: (time) => print(time),
-            totalDistanceCallback: (distance) => print(distance),
+              // demonstrates how to interact with the controller
+              Padding(
+                padding: const EdgeInsets.all(10),
+                child: Row(
+                  children: [
+                    Expanded(
+                      child: ElevatedButton(
+                        onPressed: () {
+                          mapsWidgetController.currentState!.setSourceLatLng(
+                            LatLng(
+                              40.484000837597925 * (Random().nextDouble()),
+                              -3.369978368282318,
+                            ),
+                          );
+                        },
+                        child: Text('Update source'),
+                      ),
+                    ),
+                    const SizedBox(width: 10),
+                    Expanded(
+                      child: ElevatedButton(
+                        onPressed: () async {
+                          final googleMapsCon = await mapsWidgetController
+                              .currentState!
+                              .getGoogleMapsController();
+                          googleMapsCon.showMarkerInfoWindow(
+                            MarkerIconInfo.sourceMarkerId,
+                          );
+                        },
+                        child: Text('Show source info'),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
           ),
         ),
       ),