+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-## [1.0.5] - 29/09/2022
+## [1.0.5] - 25/10/2022
 
+* Changed internal implementation of the widget
+* Added `layoutDirection` property
+* `MarkerIconInfo` inputs are now non-nullable
+* Exposed state class to allow updating source/destination lat lng, or interacting with google maps con directly
 * Updated a dependency to the latest release
+* Updated example app
+* Updated README.md
 
 ## [1.0.4] - 15/06/2022